pyproject.toml
[project]
name = "htmlcalendar"
version = "0.1.0"
description = "Generate Indonesian holidays HTML calendar"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4",
    "types-beautifulsoup4",
    "lesley",
    "requests",
    "types-requests",
    "pandas-stubs",
    "phantombrowser",
    
]

[dependency-groups]
dev = [
    "ruff",
    "mypy",
]
